Modularised Backend[B]
- Game logic is now contained in the Game object.
- Class defs are in the classes.py
- Long lists are in lists.py
