+++ backend/app.py
@@ -11,332 +11,16 @@ from flask_cors import CORS, cross_origin
 import json
 from flask_socketio import SocketIO
 
+from classes import *
+import lists
+from Game import Game
+
 app = Flask(__name__)
 cors = CORS(app)
 app.config['CORS_HEADERS'] = 'Content-Type'
 socketio = SocketIO(app, cors_allowed_origins="*")
 
 LOG_FILE = "runtime.log"
-
-# standard deck is a list of the name strings of cards
-standard_deck: list[str] = [
-    "communitysupport",
-    "communitysupport",
-    "communitysupport",
-    "communitysupport",
-
-    "event-1",
-    "event-2",
-    "event-3",
-    "event-4",
-    "event-5",
-    "event-6",
-    "event-7",
-    "event-8",
-    "event-9",
-    "event-10",
-    "event-11",
-    "event-12",
-
-    "civil-1",
-    "civil-1",
-    "civil-2",
-    "civil-2",
-    "civil-3",
-    "digital-1",
-    "digital-1",
-    "digital-2",
-    "digital-2",
-    "digital-3",
-    "economic-1",
-    "economic-2",
-    "economic-3",
-    "economic-4",
-    "economic-5",
-    "military-1",
-    "military-1",
-    "military-2",
-    "military-3",
-    "military-4",
-    "psychological-1",
-    "psychological-1",
-    "psychological-2",
-    "psychological-2",
-    "psychological-3",
-    "social-1",
-    "social-1",
-    "social-2",
-    "social-2",
-    "social-3",
-]
-
-# select crisis from here
-crisis_deck: list[str] = [
-    "crisis-1",
-    "crisis-2",
-    "crisis-3",
-    "crisis-4",
-    "crisis-5",
-    "crisis-6",
-]
-
-# get name from id string
-lookup: dict[str, str] = {
-    "communitysupport": "Community Support",
-    "event-1": "Racial and Religious Tension",
-    "event-2": "Complaint Culture",
-    "event-3": "Stolen Personal Data",
-    "event-4": "Desensitised Citizens",
-    "event-5": "Disgruntled NSmen",
-    "event-6": "Criminal Collaboration",
-    "event-7": "Blood Bank Runs Dry",
-    "event-8": "Ransomware Attack",
-    "event-9": "Heightened Terrorist Alert",
-    "event-10": "Self Radicalisation",
-    "event-11": "Lone Wolf Attack",
-    "event-12": "Recruiting Desperate People",
-    "civil-1": "Alert Community",
-    "civil-2": "Prepared for Crisis",
-    "civil-3": "Civil Emergency",
-    "digital-1": "Use of Strong Passwords",
-    "digital-2": "Robust Emergency Protocols",
-    "digital-3": "Responsible Social Media Use",
-    "economic-1": "Career Mobility",
-    "economic-2": "Business Resilience",
-    "economic-3": "Financial Aid Schemes",
-    "economic-4": "Increase Surveillance at Maritime Ports",
-    "economic-5": "Skills Upgrading",
-    "military-1": "NSmen on Guard",
-    "military-2": "Increased Security Checks",
-    "military-3": "Naval Convoys",
-    "military-4": "Conduct of Raids",
-    "psychological-1": "Education",
-    "psychological-2": "Attend National Day Celebrations",
-    "psychological-3": "Strong Resolve",
-    "social-1": "Religious Counselling",
-    "social-2": "Neighbourliness",
-    "social-3": "Show of Solidarity"
-}
-
-
-class Player:
-    def __init__(self, name):
-        self.name = name
-        self.deck: list[str] = standard_deck.copy()
-        random.shuffle(self.deck)  # upon creation of the user, shuffle the deck
-        self.crisis: str = random.choice(crisis_deck)  # server side crisis
-        self.hand: list[dict[str, Union[str, bool]]] = []  # server side hand
-        self.discard: list[str] = []
-        self.field: list[str] = []
-        self.storage = {
-            "showDialogNormal": False,
-            "showDialogDefence": False,
-            "showOptionDefence": False,
-            "selectionDefence": [],
-            "showOptionDefence2": False,
-            "showOptionField": False,
-            "showDialogField": False,
-            "showDiscardPlay": False,
-            "showOptionHand": False,
-            "showDialogHand": False,
-            "opponentHandTemp": [],
-            "discardHand": False,
-            "canClickEndTurn": True,
-            "index": -1
-        }
-
-    def shuffleDeck(self):
-        random.shuffle(self.deck)
-
-    def popDeck(self):
-        return self.deck.pop()
-
-    def newDeck(self):
-        self.deck = standard_deck.copy()
-        self.discard = []
-        self.hand = []
-        random.shuffle(self.deck)
-
-    def newCrisis(self):
-        self.crisis = random.choice(crisis_deck)
-        return self.crisis
-
-    def setHandEnablePlayStatus(self, status: bool):
-        handcopy = self.hand.copy()
-        for i in handcopy:
-            i["enablePlay"] = status
-        self.hand = handcopy
-        return self.hand
-
-    def addHandCard(self, cardName):
-        self.hand.append({"name": cardName, "enablePlay": True,  # "blockPlay": False, # can play card without effect
-                          # these will be set later
-                          # everywhere that could affect this (e.g. draw card) will be accompanied by a recompute call
-                          # we just cannot call it here because this is Player not Game
-                          "requiresDialogNormal": False, "requiresOptionDefence": False,
-                          "requiresOptionField": False, "requiresDialogHand": False,
-                          "requiresOptionHand": False})
-        return self.hand
-
-    def defenceCheck(self):
-        # check defences
-        # will return a string like "CDEMPS" where a letter appears if it is in the field
-        # and does not appear if it is not
-        # ignores community support
-        return "".join(list(sorted(set([i[0].upper() for i in self.field if i != "communitysupport"]))))
-
-    def gameDefenceFulfilled(self):
-        if self.defenceCheck() == "CDEMPS":
-            return 5
-        elif self.crisis == "crisis-1":
-            return len(self.defenceCheck().replace("M", ""))
-        elif self.crisis == "crisis-2":
-            return len(self.defenceCheck().replace("C", ""))
-        elif self.crisis == "crisis-3":
-            return len(self.defenceCheck().replace("E", ""))
-        elif self.crisis == "crisis-4":
-            return len(self.defenceCheck().replace("S", ""))
-        elif self.crisis == "crisis-5":
-            return len(self.defenceCheck().replace("P", ""))
-        elif self.crisis == "crisis-6":
-            return len(self.defenceCheck().replace("D", ""))
-
-    def gameWon(self):
-        return self.gameDefenceFulfilled() == 5
-
-
-class Game:
-    def __init__(self, player1_username, player2_username, internal_id):
-        self.player1 = Player(player1_username)
-        self.player2 = Player(player2_username)
-        while self.player1.crisis == self.player2.crisis:
-            # crises cannot be equal
-            self.player1 = Player(player1_username)
-            self.player2 = Player(player2_username)
-        self.player1_username = player1_username
-        self.player2_username = player2_username
-        self.gofirst = self.player1_username if self.player1.crisis > self.player2.crisis else self.player2_username
-        self.turn = self.gofirst
-        self.nextturnislast = False
-
-        self.internal_id = internal_id
-
-    def recomputeBlockAndDialogStatus(self):
-        handcopy = self.player1.hand.copy()
-        for i in handcopy:
-            # [1 CSC:] Draw 2 additional cards OR Take any card from your discard pile and place it back into your hand
-            if i["name"] in ("military-2", "military-3", "civil-2", "economic-3", "economic-4"):
-                i["requiresDialogNormal"] = self.player1.field.count("communitysupport") >= 1
-                if len(self.player1.discard) == 0:
-                    i[
-                        "warn"] = "\nWarning: There are no cards in your discard pile. Picking the second option will have no effect."
-                else:
-                    i["warn"] = ""  # no warning for playing def card without effect as it could be helpful
-                    # for example to win the game
-            # If your opponent has 1 or less Community Support points, discard any 2 Defence cards from your opponent's field
-            if i["name"] in ("event-5", "event-6", "event-7", "event-8"):
-                i["requiresOptionDefence"] = True
-                if self.player2.field.count("communitysupport") <= 1 and (
-                        len(self.player2.field) - self.player2.field.count("communitysupport") >= 2
-                ):
-                    i["warn"] = ""
-                elif self.player1.field.count("communitysupport") <= 1 and (
-                        len(self.player2.field) - self.player2.field.count("communitysupport") == 1
-                ):
-                    i["warn"] = "\nWarning: Opponent has only 1 defence card to select."
-                elif self.player1.field.count("communitysupport") <= 1 and (
-                        len(self.player2.field) - self.player2.field.count("communitysupport") == 0
-                ):
-                    i["warn"] = "\nWarning: Opponent has no defence cards to select. This card will have no effect."
-                else:
-                    i["warn"] = "\nWarning: Opponent has >1 community support. This card will have no effect."
-            # If your opponent has 2 or less Community Support points, discard any 1 card from your opponent's field
-            if i["name"] in ("event-9", "event-10", "event-11", "event-12"):
-                i["requiresOptionField"] = True
-                if self.player2.field.count("communitysupport") <= 2 and len(self.player2.field) >= 1:
-                    i["warn"] = ""
-                elif self.player2.field.count("communitysupport") <= 2 and len(self.player2.field) == 0:
-                    i["warn"] = "\nWarning: Opponent has no field cards to select. This card will have no effect."
-                else:
-                    i["warn"] = "\nWarning: Opponent has >2 community support. This card will have no effect."
-            # If your opponent has no Community Support points, look at your opponent's hand and discard 1 card from there
-            if i["name"] in ("event-1", "event-3", "event-2", "event-4"):
-                i["requiresOptionHand"] = True
-                if len(self.player2.hand) > 0 and self.player2.field.count("communitysupport") == 0:
-                    i["warn"] = ""
-                elif len(self.player2.hand) == 0 and self.player2.field.count("communitysupport") == 0:
-                    i["warn"] = "\nWarning: Opponent has no hand cards to select. This card will have no effect."
-                else:
-                    i["warn"] = "\nWarning: Opponent has >0 community support. This card will have no effect!"
-            # Look at the cards in your opponent's hand. [2 CSC:] Draw 1 additional card AND play 1 additional card this turn
-            if i["name"] in ("military-4", "civil-3", "economic-5", "social-3", "psychological-3", "digital-3"):
-                i["requiresDialogHand"] = True
-
-        self.player1.hand = handcopy
-
-        handcopy = self.player2.hand.copy()
-        for i in handcopy:
-            # [1 CSC:] Draw 2 additional cards OR Take any card from your discard pile and place it back into your hand
-            if i["name"] in ("military-2", "military-3", "civil-2", "economic-3", "economic-4"):
-                i["requiresDialogNormal"] = self.player2.field.count("communitysupport") >= 1
-                if len(self.player2.discard) == 0:
-                    i[
-                        "warn"] = "\nWarning: There are no cards in your discard pile. Picking the second option will have no effect."
-                else:
-                    i["warn"] = ""  # no warning for playing def card without effect as it could be helpful
-                    # for example to win the game
-            # If your opponent has 1 or less Community Support points, discard any 2 Defence cards from your opponent's field
-            if i["name"] in ( "event-5", "event-6", "event-7", "event-8"):
-                i["requiresOptionDefence"] = True
-                if self.player1.field.count("communitysupport") <= 1 and (
-                        len(self.player1.field) - self.player1.field.count("communitysupport") >= 2
-                ):
-                    i["warn"] = ""
-                elif self.player1.field.count("communitysupport") <= 1 and (
-                        len(self.player1.field) - self.player1.field.count("communitysupport") == 1
-                ):
-                    i["warn"] = "\nWarning: Opponent has only 1 defence card to select."
-                elif self.player1.field.count("communitysupport") <= 1 and (
-                        len(self.player1.field) - self.player1.field.count("communitysupport") == 0
-                ):
-                    i["warn"] = "\nWarning: Opponent has no defence cards to select. This card will have no effect."
-                else:
-                    i["warn"] = "\nWarning: Opponent has >1 community support. This card will have no effect."
-            # If your opponent has 2 or less Community Support points, discard any 1 card from your opponent's field
-            if i["name"] in ("event-9", "event-10", "event-11", "event-12"):
-                i["requiresOptionField"] = True
-                if self.player1.field.count("communitysupport") <= 2 and len(self.player1.field) >= 1:
-                    i["warn"] = ""
-                elif self.player1.field.count("communitysupport") <= 2 and len(self.player1.field) == 0:
-                    i["warn"] = "\nWarning: Opponent has no field cards to select. This card will have no effect."
-                else:
-                    i["warn"] = "\nWarning: Opponent has >2 community support. This card will have no effect."
-            # If your opponent has no Community Support points, look at your opponent's hand and discard 1 card from there
-            if i["name"] in ("event-1", "event-3", "event-2", "event-4"):
-                i["requiresOptionHand"] = True
-                if len(self.player1.hand) > 0 and self.player1.field.count("communitysupport") == 0:
-                    i["warn"] = ""
-                elif len(self.player1.hand) == 0 and self.player1.field.count("communitysupport") == 0:
-                    i["warn"] = "\nWarning: Opponent has no hand cards to select. This card will have no effect."
-                else:
-                    i["warn"] = "\nWarning: Opponent has >0 community support. This card will have no effect!"
-            # Look at the cards in your opponent's hand. [2 CSC:] Draw 1 additional card AND play 1 additional card this turn
-            if i["name"] in ("military-4", "civil-3", "economic-5", "social-3", "psychological-3", "digital-3"):
-                i["requiresDialogHand"] = True
-
-        self.player2.hand = handcopy
-        return self.player1.hand, self.player2.hand
-
-
-class User:
-    def __init__(self, name, last_checkin, login_session_key):
-        self.name: str = name
-        self.last_checkin: float = last_checkin
-        self.login_session_key: str = login_session_key
-        self.games: list[str] = []
-
-    def __str__(self):
-        return self.name + " | " + str(self.last_checkin)
 
 
 logged_in: list[User] = []
@@ -895,7 +579,7 @@ def new_crisis():
 @app.route('/deck_size')
 @cross_origin()
 def new_deck_size():
-    return str(len(standard_deck))
+    return str(len(lists.STANDARD_DECK))
 
 
 @app.route('/get_crisis', methods=["POST"])
@@ -997,501 +681,15 @@ def play_hand():
             keys_equal = secrets.compare_digest(i.login_session_key, sent_login_sesh_key)
             if (i.name == your_username) and keys_equal:
                 game: Game = games[game_id]
-                if game.player1_username == request_username:
-                    if (hand_index < 0) or (hand_index >= len(game.player1.hand)):
-                        abort(Response(json.dumps({"Message": "Card Index Out Of Range"}), 422))
-                    else:
-                        cardPlayed = game.player1.hand.pop(hand_index)["name"]
-
-                        if cardPlayed == "communitysupport":
-                            game.player1.field.append("communitysupport")
-                            your_move_notifier = "You played Community Support."
-                            opponent_move_notifier = "Opponent played Community Support."
-                            next_turn = True
-                        elif cardPlayed in ("military-1", "economic-1", "economic-2", "civil-1"):  # draw 1 card
-                            game.player1.field.append(cardPlayed)
-                            if len(game.player1.deck) > 0:
-                                poppedCard = game.player1.popDeck()
-                                game.player1.addHandCard(poppedCard)  # add the popped card to the hand
-                                your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 1 card."
-                            else:
-                                your_move_notifier = f"You played {lookup[cardPlayed]} as you only had 0 cards in your deck."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} as they had only 0 cards in their deck."
-                            next_turn = True
-                        elif cardPlayed in ("military-2", "military-3", "civil-2", "economic-3", "economic-4"):
-                            game.player1.field.append(cardPlayed)
-                            if "extra" in request.json:
-                                extra = request.json["extra"]  # -1 if draw 2, else restore nth discarded card
-                                if extra == -1:
-                                    if len(game.player1.deck) > 1:
-                                        poppedCard = game.player1.popDeck()
-                                        game.player1.addHandCard(poppedCard)  # add the popped card to the hand
-                                        poppedCard2 = game.player1.popDeck()
-                                        game.player1.addHandCard(poppedCard2)  # add the popped card to the hand
-                                        your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]}, {lookup[poppedCard2]}."
-                                        opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 2 cards."
-                                    elif len(game.player1.deck) == 1:
-                                        poppedCard = game.player1.popDeck()
-                                        game.player1.addHandCard(poppedCard)  # add the popped card to the hand
-                                        your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]} as you only had 1 card in your deck."
-                                        opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 1 card as they had only 1 card in their deck."
-                                    else:
-                                        your_move_notifier = f"You played {lookup[cardPlayed]} as you only had 0 cards in your deck."
-                                        opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} as they had only 0 cards in their deck."
-                                else:
-                                    restore = game.player1.discard.pop(extra)
-                                    game.player1.addHandCard(restore)
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} and restored {lookup[restore]}."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and restored {lookup[restore]}."
-                            else:
-                                # no additional effect
-                                your_move_notifier = f"You played {lookup[cardPlayed]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}."
-                            next_turn = True
-                        elif cardPlayed in (
-                                "psychological-2", "social-2",
-                                "digital-2"):  # draw 1 card, 1 extra if have community support
-                            game.player1.field.append(cardPlayed)
-                            if game.player1.field.count("communitysupport") >= 1:
-                                if len(game.player1.deck) > 1:
-                                    poppedCard = game.player1.popDeck()
-                                    game.player1.addHandCard(poppedCard)  # add the popped card to the hand
-                                    poppedCard2 = game.player1.popDeck()
-                                    game.player1.addHandCard(poppedCard2)  # add the popped card to the hand
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]}, {lookup[poppedCard2]}."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 2 cards."
-                                elif len(game.player1.deck) == 1:
-                                    poppedCard = game.player1.popDeck()
-                                    game.player1.addHandCard(poppedCard)  # add the popped card to the hand
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]} as you only had 1 card in your deck."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 1 card as they had only 1 card in their deck."
-                                else:
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} as you only had 0 cards in your deck."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} as they had only 0 cards in their deck."
-                            else:
-                                if len(game.player1.deck) > 0:
-                                    poppedCard = game.player1.popDeck()
-                                    game.player1.addHandCard(poppedCard)  # add the popped card to the hand
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]}."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 1 card."
-                                else:
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} as you only had 0 cards in your deck."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} as they had only 0 cards in their deck."
-                            next_turn = True
-                        elif cardPlayed in ("social-1", "psychological-1", "digital-1"):
-                            game.player1.field.append(cardPlayed)
-                            your_move_notifier = f"You played {lookup[cardPlayed]}. It is still your turn."
-                            opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}. It is still their turn."
-                            next_turn = False
-                        elif cardPlayed in (
-                                "military-4", "civil-3", "economic-5", "social-3", "psychological-3", "digital-3"):
-                            game.player1.field.append(cardPlayed)
-                            if game.player1.field.count("communitysupport") >= 2:
-                                # draw 1, extra turn
-                                if len(game.player1.deck) > 0:
-                                    poppedCard = game.player1.popDeck()
-                                    game.player1.addHandCard(poppedCard)  # add the popped card to the hand
-                                    your_move_notifier = f"You played {lookup[cardPlayed]}, viewed your opponent's hand and drew {lookup[poppedCard]}. It is still your turn."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}, viewed your hand and drew 1 card. It is still their turn."
-                                else:
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} as you only had 0 cards in your deck. It is still your turn."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} as they had only 0 cards in their deck. It is still their turn."
-                                next_turn = False
-                            else:
-                                your_move_notifier = f"You played {lookup[cardPlayed]} and viewed your opponent's hand."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and viewed your hand."
-                                next_turn = True
-                        elif cardPlayed in ("event-1", "event-3", "event-4"):
-                            game.player1.discard = [cardPlayed] + game.player1.discard
-                            if "extra" in request.json:
-                                extra = request.json["extra"]
-                                card1 = game.player2.hand.pop(extra)["name"]
-                                game.player2.discard = [card1] + game.player2.discard
-                                your_move_notifier = f"You played {lookup[cardPlayed]}, discarding opponent's {lookup[card1]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}, discarding your {lookup[card1]}."
-                            else:
-                                # no additional effect
-                                your_move_notifier = f"You played {lookup[cardPlayed]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}."
-                            next_turn = True
-                        elif cardPlayed in ("event-2", "event-5", "event-6", "event-7", "event-8"):
-                            game.player1.discard = [cardPlayed] + game.player1.discard
-                            if "extra" in request.json:
-                                extra = request.json["extra"]  # [which, key]
-                                card1 = list(filter(lambda name: name.startswith(extra[0]), game.player2.field))[
-                                    extra[1]]
-                                game.player2.discard = [card1] + game.player2.discard
-                                if "extra2" in request.json:
-                                    extra2 = request.json["extra2"]
-                                    card2 = list(filter(lambda name: name.startswith(extra2[0]), game.player2.field))[
-                                        extra2[1]]
-                                    # todo (low priority) preserve order
-                                    game.player2.field.remove(card1)
-                                    game.player2.field.remove(card2)
-                                    game.player2.discard = [card2] + game.player2.discard
-                                    your_move_notifier = f"You played {lookup[cardPlayed]}, discarding opponent's {lookup[card1]} & {lookup[card2]}."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}, discarding your {lookup[card1]} & {lookup[card2]}."
-                                else:
-                                    game.player2.field.remove(card1)
-                                    your_move_notifier = f"You played {lookup[cardPlayed]}, discarding opponent's {lookup[card1]}."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}, discarding your {lookup[card1]}."
-                            else:
-                                # no additional effect
-                                your_move_notifier = f"You played {lookup[cardPlayed]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}."
-                            next_turn = True
-                        elif cardPlayed in ("event-9", "event-10", "event-11", "event-12"):
-                            game.player1.discard = [cardPlayed] + game.player1.discard
-                            if "extra" in request.json:
-                                extra = request.json["extra"]  # [which, key]
-                                card1 = list(filter(lambda name: name.startswith(extra[0]), game.player2.field))[
-                                    extra[1]]
-                                game.player2.discard = [card1] + game.player2.discard
-                                game.player2.field.remove(card1)
-                                your_move_notifier = f"You played {lookup[cardPlayed]}, discarding opponent's {lookup[card1]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}, discarding your {lookup[card1]}."
-                            else:
-                                # no additional effect
-                                your_move_notifier = f"You played {lookup[cardPlayed]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}."
-                            next_turn = True
-                        else:  # UNUSED
-                            game.player1.discard = [cardPlayed] + game.player1.discard
-                            your_move_notifier = f"You discarded {lookup[cardPlayed]}."
-                            opponent_move_notifier = f"Opponent discarded {lookup[cardPlayed]}."
-                            next_turn = False
-
-                        if game.player1.gameWon():
-                            your_move_notifier += "\nYou win!"
-                            opponent_move_notifier += "\nYou lose!"
-                            next_turn = False
-                            game.player1.setHandEnablePlayStatus(False)
-                            game.player2.setHandEnablePlayStatus(False)
-                            response["winThisTurn"] = True
-
-                        if next_turn:
-                            if len(game.player1.hand) > 7:
-                                # the max cards you can have is 9. 7 cards, draw 1 for turn, play 1 allowing draw 2.
-                                your_move_notifier += f"\nPlease discard cards until you have 7 cards."
-                                opponent_move_notifier += f"\nWaiting for opponent to discard hand cards."
-
-                                response["needDiscard"] = True
-                            else:
-                                if game.nextturnislast:
-                                    if game.player1.gameDefenceFulfilled() > game.player2.gameDefenceFulfilled():
-                                        your_move_notifier += "\nOpponent's deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                                        opponent_move_notifier += "\nYour deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                                    elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                        your_move_notifier += "\nOpponent's deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                                        opponent_move_notifier += "\nYour deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                                    else:  # player 2 more than player 1
-                                        your_move_notifier += "\nYOpponent's deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                                        opponent_move_notifier += "\nYour deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                                    response["winThisTurn"] = True
-                                elif len(game.player1.deck) == 0:
-                                    if game.gofirst == game.player1_username:
-                                        # award extra turn
-                                        game.nextturnislast = True
-                                        your_move_notifier += "\nWarning: Your opponent's next turn is the last turn, due to your deck running out of cards."
-                                        opponent_move_notifier += "\nWarning: Your opponent's next turn is the last, due to their deck running out of cards."
-                                    else:
-                                        # end game and compute winner
-                                        if game.player1.gameDefenceFulfilled() > game.player2.gameDefenceFulfilled():
-                                            your_move_notifier += "\nYour deck ran out of cards. You win due to having more defence fulfilled!"
-                                            opponent_move_notifier += "\nOpponent's deck ran out of cards. You lose due to having less defence fulfilled!"
-                                        elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                            your_move_notifier += "\nYour deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                            opponent_move_notifier += "\nOpponent's deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                        else:  # player 2 more than player 1
-                                            your_move_notifier += "\nYour deck ran out of cards. You lose due to having less defence fulfilled!"
-                                            opponent_move_notifier += "\nOpponent's deck ran out of cards. You win due to having more defence fulfilled!"
-                                        response["winThisTurn"] = True
-                                else:
-                                    poppedCard = game.player2.popDeck()
-                                    game.player2.addHandCard(poppedCard)  # add the popped card to the hand
-
-                                    game.player2.setHandEnablePlayStatus(True)
-                                    game.player1.setHandEnablePlayStatus(False)
-                                    game.recomputeBlockAndDialogStatus()
-                                    game.turn = game.player2_username
-
-                                    your_move_notifier += "\nOpponent's turn."
-                                    opponent_move_notifier += f"\nYour turn. You drew {lookup[poppedCard]}."
-
-                        response["hand"] = game.player1.hand
-                        response["discard"] = game.player1.discard
-                        response["cardsLeft"] = len(game.player1.deck)
-                        response["cardPlayed"] = cardPlayed
-                        response["field"] = game.player1.field
-                        response["moveNotifier"] = your_move_notifier
-                        response["canClickEndTurn"] = game.player1_username == game.turn
-
-                        updater = {"uuid": game_id, "username": game.player2_username,
-                                   "hand": game.player1.hand, "discard": game.player1.discard,
-                                   "cardsLeft": len(game.player1.deck), "field": game.player1.field,
-                                   "moveNotifier": opponent_move_notifier}
-                        socketio.emit("update opponent state", updater)
-
-                        updater = {"uuid": game_id, "username": game.player2_username,
-                                   "hand": game.player2.hand, "discard": game.player2.discard,
-                                   "cardsLeft": len(game.player2.deck), "field": game.player2.field,
-                                   "canClickEndTurn": game.turn == game.player2_username}
-                        socketio.emit("update your state", updater)
-
-                        updater = {"uuid": game_id, "username": game.player1_username,
-                                   "hand": game.player2.hand, "discard": game.player2.discard,
-                                   "cardsLeft": len(game.player2.deck), "field": game.player2.field}
-                        socketio.emit("update opponent state", updater)
-
-                        return response
-                elif game.player2_username == request_username:
-                    if (hand_index < 0) or (hand_index >= len(game.player2.hand)):
-                        abort(Response(json.dumps({"Message": "Card Index Out Of Range"}), 422))
-                    else:
-                        cardPlayed = game.player2.hand.pop(hand_index)["name"]
-
-                        if cardPlayed == "communitysupport":
-                            game.player2.field.append("communitysupport")
-                            your_move_notifier = "You played Community Support."
-                            opponent_move_notifier = "Opponent played Community Support."
-                            next_turn = True
-                        elif cardPlayed in ("military-1", "economic-1", "economic-2", "civil-1"):  # draw 1 card
-                            game.player2.field.append(cardPlayed)
-                            if len(game.player2.deck) > 0:
-                                poppedCard = game.player2.popDeck()
-                                game.player2.addHandCard(poppedCard)  # add the popped card to the hand
-                                your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 1 card."
-                            else:
-                                your_move_notifier = f"You played {lookup[cardPlayed]} as you only had 0 cards in your deck."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} as they had only 0 cards in their deck."
-                            next_turn = True
-                        elif cardPlayed in ("military-2", "military-3", "civil-2", "economic-3", "economic-4"):
-                            game.player2.field.append(cardPlayed)
-                            if "extra" in request.json:
-                                extra = request.json["extra"]  # -1 if draw 2, else restore nth discarded card
-                                if extra == -1:
-                                    if len(game.player2.deck) > 1:
-                                        poppedCard = game.player2.popDeck()
-                                        game.player2.addHandCard(poppedCard)  # add the popped card to the hand
-                                        poppedCard2 = game.player2.popDeck()
-                                        game.player2.addHandCard(poppedCard2)  # add the popped card to the hand
-                                        your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]}, {lookup[poppedCard2]}."
-                                        opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 2 cards."
-                                    elif len(game.player2.deck) == 1:
-                                        poppedCard = game.player2.popDeck()
-                                        game.player2.addHandCard(poppedCard)  # add the popped card to the hand
-                                        your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]} as you only had 1 card in your deck."
-                                        opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 1 card as they had only 1 card in their deck."
-                                    else:
-                                        your_move_notifier = f"You played {lookup[cardPlayed]} as you only had 0 cards in your deck."
-                                        opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} as they had only 0 cards in their deck."
-                                else:
-                                    restore = game.player2.discard.pop(extra)
-                                    game.player2.addHandCard(restore)
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} and restored {lookup[restore]}."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and restored {lookup[restore]}."
-                            else:
-                                # no additional effect
-                                your_move_notifier = f"You played {lookup[cardPlayed]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}."
-                            next_turn = True
-                        elif cardPlayed in (
-                                "psychological-2", "social-2",
-                                "digital-2"):  # draw 1 card, 1 extra if have community support
-                            game.player2.field.append(cardPlayed)
-                            if game.player2.field.count("communitysupport") >= 1:
-                                if len(game.player2.deck) > 1:
-                                    poppedCard = game.player2.popDeck()
-                                    game.player2.addHandCard(poppedCard)  # add the popped card to the hand
-                                    poppedCard2 = game.player2.popDeck()
-                                    game.player2.addHandCard(poppedCard2)  # add the popped card to the hand
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]}, {lookup[poppedCard2]}."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 2 cards."
-                                elif len(game.player2.deck) == 1:
-                                    poppedCard = game.player2.popDeck()
-                                    game.player2.addHandCard(poppedCard)  # add the popped card to the hand
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]} as you only had 1 card in your deck."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 1 card as they had only 1 card in their deck."
-                                else:
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} as you only had 0 cards in your deck."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} as they had only 0 cards in their deck."
-                            else:
-                                if len(game.player2.deck) > 0:
-                                    poppedCard = game.player2.popDeck()
-                                    game.player2.addHandCard(poppedCard)  # add the popped card to the hand
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} and drew {lookup[poppedCard]}."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and drew 1 card."
-                                else:
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} as you only had 0 cards in your deck."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} as they had only 0 cards in their deck."
-                            next_turn = True
-                        elif cardPlayed in ("social-1", "psychological-1", "digital-1"):
-                            game.player2.field.append(cardPlayed)
-                            your_move_notifier = f"You played {lookup[cardPlayed]}. It is still your turn."
-                            opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}. It is still their turn."
-                            next_turn = False
-                        elif cardPlayed in (
-                                "military-4", "civil-3", "economic-5", "social-3", "psychological-3", "digital-3"):
-                            game.player2.field.append(cardPlayed)
-                            if game.player2.field.count("communitysupport") >= 2:
-                                # draw 1, extra turn
-                                if len(game.player2.deck) > 0:
-                                    poppedCard = game.player2.popDeck()
-                                    game.player2.addHandCard(poppedCard)  # add the popped card to the hand
-                                    your_move_notifier = f"You played {lookup[cardPlayed]}, viewed your opponent's hand and drew {lookup[poppedCard]}. It is still your turn."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}, viewed your hand and drew 1 card. It is still their turn."
-                                else:
-                                    your_move_notifier = f"You played {lookup[cardPlayed]} as you only had 0 cards in your deck. It is still your turn."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} as they had only 0 cards in their deck. It is still their turn."
-                                next_turn = False
-                            else:
-                                your_move_notifier = f"You played {lookup[cardPlayed]} and viewed your opponent's hand."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]} and viewed your hand."
-                                next_turn = True
-                        elif cardPlayed in ("event-1", "event-3", "event-4"):
-                            game.player2.discard = [cardPlayed] + game.player2.discard
-                            if "extra" in request.json:
-                                extra = request.json["extra"]
-                                card1 = game.player1.hand.pop(extra)["name"]
-                                game.player1.discard = [card1] + game.player1.discard
-                                your_move_notifier = f"You played {lookup[cardPlayed]}, discarding opponent's {lookup[card1]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}, discarding your {lookup[card1]}."
-                            else:
-                                # no additional effect
-                                your_move_notifier = f"You played {lookup[cardPlayed]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}."
-                            next_turn = True
-                        elif cardPlayed in ("event-2", "event-5", "event-6", "event-7", "event-8"):
-                            game.player2.discard = [cardPlayed] + game.player2.discard
-                            if "extra" in request.json:
-                                extra = request.json["extra"]  # [which, key]
-                                card1 = list(filter(lambda name: name.startswith(extra[0]), game.player1.field))[
-                                    extra[1]]
-                                game.player1.discard = [card1] + game.player1.discard
-                                if "extra2" in request.json:
-                                    extra2 = request.json["extra2"]
-                                    card2 = list(filter(lambda name: name.startswith(extra2[0]), game.player1.field))[
-                                        extra2[1]]
-                                    # todo (low priority) preserve order
-                                    game.player1.field.remove(card1)
-                                    game.player1.field.remove(card2)
-                                    game.player1.discard = [card2] + game.player1.discard
-                                    your_move_notifier = f"You played {lookup[cardPlayed]}, discarding opponent's {lookup[card1]} & {lookup[card2]}."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}, discarding your {lookup[card1]} & {lookup[card2]}."
-                                else:
-                                    game.player1.field.remove(card1)
-                                    your_move_notifier = f"You played {lookup[cardPlayed]}, discarding opponent's {lookup[card1]}."
-                                    opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}, discarding your {lookup[card1]}."
-                            else:
-                                # no additional effect
-                                your_move_notifier = f"You played {lookup[cardPlayed]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}."
-                            next_turn = True
-                        elif cardPlayed in ("event-9", "event-10", "event-11", "event-12"):
-                            game.player2.discard = [cardPlayed] + game.player2.discard
-                            if "extra" in request.json:
-                                extra = request.json["extra"]  # [which, key]
-                                card1 = list(filter(lambda name: name.startswith(extra[0]), game.player1.field))[
-                                    extra[1]]
-                                game.player1.discard = [card1] + game.player1.discard
-                                game.player1.field.remove(card1)
-                                your_move_notifier = f"You played {lookup[cardPlayed]}, discarding opponent's {lookup[card1]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}, discarding your {lookup[card1]}."
-                            else:
-                                # no additional effect
-                                your_move_notifier = f"You played {lookup[cardPlayed]}."
-                                opponent_move_notifier = f"Opponent played {lookup[cardPlayed]}."
-                            next_turn = True
-                        else:  # UNUSED
-                            game.player2.discard = [cardPlayed] + game.player2.discard
-                            your_move_notifier = f"You discarded {lookup[cardPlayed]}."
-                            opponent_move_notifier = f"Opponent discarded {lookup[cardPlayed]}."
-                            next_turn = False
-
-                        if game.player2.gameWon():
-                            your_move_notifier += "\nYou win!"
-                            opponent_move_notifier += "\nYou lose!"
-                            next_turn = False
-                            game.player1.setHandEnablePlayStatus(False)
-                            game.player2.setHandEnablePlayStatus(False)
-                            response["winThisTurn"] = True
-
-                        if next_turn:
-                            if len(game.player2.hand) > 7:
-                                # the max cards you can have is 9. 7 cards, draw 1 for turn, play 1 allowing draw 2.
-                                your_move_notifier += f"\nPlease discard cards until you have 7 cards."
-                                opponent_move_notifier += f"\nWaiting for opponent to discard hand cards."
-                                response["needDiscard"] = True
-                            else:
-                                if game.nextturnislast:
-                                    if game.player1.gameDefenceFulfilled() < game.player2.gameDefenceFulfilled():
-                                        your_move_notifier += "\nOpponent's deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                                        opponent_move_notifier += "\nYour deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                                    elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                        your_move_notifier += "\nOpponent's deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                                        opponent_move_notifier += "\nYour deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                                    else:  # player 2 more than player 1
-                                        your_move_notifier += "\nYOpponent's deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                                        opponent_move_notifier += "\nYour deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                                    response["winThisTurn"] = True
-                                elif len(game.player2.deck) == 0:
-                                    if game.gofirst == game.player2_username:
-                                        # award extra turn
-                                        game.nextturnislast = True
-                                        your_move_notifier += "\nWarning: Your opponent's next turn is the last turn, due to your deck running out of cards."
-                                        opponent_move_notifier += "\nWarning: Your opponent's next turn is the last, due to their deck running out of cards."
-                                    else:
-                                        # end game and compute winner
-                                        if game.player1.gameDefenceFulfilled() < game.player2.gameDefenceFulfilled():
-                                            your_move_notifier += "\nYour deck ran out of cards. You win due to having more defence fulfilled!"
-                                            opponent_move_notifier += "\nOpponent's deck ran out of cards. You lose due to having less defence fulfilled!"
-                                        elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                            your_move_notifier += "\nYour deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                            opponent_move_notifier += "\nOpponent's deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                        else:  # player 2 more than player 1
-                                            your_move_notifier += "\nYour deck ran out of cards. You lose due to having less defence fulfilled!"
-                                            opponent_move_notifier += "\nOpponent's deck ran out of cards. You win due to having more defence fulfilled!"
-                                        response["winThisTurn"] = True
-                                else:
-                                    poppedCard = game.player1.popDeck()
-                                    game.player1.addHandCard(poppedCard)  # add the popped card to the hand
-
-                                    game.player1.setHandEnablePlayStatus(True)
-                                    game.player2.setHandEnablePlayStatus(False)
-                                    game.recomputeBlockAndDialogStatus()
-                                    game.turn = game.player1_username
-
-                                    your_move_notifier += "\nOpponent's turn."
-                                    opponent_move_notifier += f"\nYour turn. You drew {lookup[poppedCard]}."
-
-                        response["hand"] = game.player2.hand
-                        response["discard"] = game.player2.discard
-                        response["cardsLeft"] = len(game.player2.deck)
-                        response["cardPlayed"] = cardPlayed
-                        response["field"] = game.player2.field
-                        response["moveNotifier"] = your_move_notifier
-                        response["canClickEndTurn"] = game.player2_username == game.turn
-
-                        updater = {"uuid": game_id, "username": game.player1_username,
-                                   "hand": game.player2.hand, "discard": game.player2.discard,
-                                   "cardsLeft": len(game.player2.deck), "field": game.player2.field,
-                                   "moveNotifier": opponent_move_notifier}
-                        socketio.emit("update opponent state", updater)
-
-                        updater = {"uuid": game_id, "username": game.player1_username,
-                                   "hand": game.player1.hand, "discard": game.player1.discard,
-                                   "cardsLeft": len(game.player1.deck), "field": game.player1.field,
-                                   "canClickEndTurn": game.turn == game.player1_username}
-                        socketio.emit("update your state", updater)
-
-                        updater = {"uuid": game_id, "username": game.player2_username,
-                                   "hand": game.player1.hand, "discard": game.player1.discard,
-                                   "cardsLeft": len(game.player1.deck), "field": game.player1.field}
-                        socketio.emit("update opponent state", updater)
-
-                        return response
+                if (
+                        (game.player1_username == request_username) and
+                        ((hand_index < 0) or (hand_index >= len(game.player1.hand)))
+                ) or (
+                        (game.player2_username == request_username) and
+                        ((hand_index < 0) or (hand_index >= len(game.player2.hand)))
+                ):
+                    abort(Response(json.dumps({"Message": "Card Index Out Of Range"}), 422))
+                return game.play_hand(socketio, 1 if game.player1_username == request_username else 2, hand_index, request)
         # else
         abort(Response(json.dumps({"Message": "Cannot Play Hand"}), 404))
 
@@ -1518,170 +716,16 @@ def discard_hand():
             keys_equal = secrets.compare_digest(i.login_session_key, sent_login_sesh_key)
             if (i.name == your_username) and keys_equal:
                 game: Game = games[game_id]
-                if game.player1_username == request_username:
-                    if (hand_index < 0) or (hand_index >= len(game.player1.hand)):
+                if (
+                    (game.player1_username == request_username) and
+                    ((hand_index < 0) or (hand_index >= len(game.player1.hand)))
+                ) or (
+                    (game.player2_username == request_username) and
+                    ((hand_index < 0) or (hand_index >= len(game.player2.hand)))
+                ):
                         abort(Response(json.dumps({"Message": "Card Index Out Of Range"}), 422))
-                    else:
-                        cardPlayed = game.player1.hand.pop(hand_index)["name"]
-                        game.player1.discard = [cardPlayed] + game.player1.discard
+                return game.discard_hand(socketio, 1 if game.player1_username == request_username else 2, hand_index)
 
-                        next_turn = len(game.player1.hand) > 7
-
-                        if next_turn:
-                            your_move_notifier = "Please discard cards until you have 7 cards."
-                            opponent_move_notifier = "Waiting for opponent to discard hand cards."
-                        else:
-                            if game.nextturnislast:
-                                if game.player1.gameDefenceFulfilled() > game.player2.gameDefenceFulfilled():
-                                    your_move_notifier = "Opponent's deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                                    opponent_move_notifier = "Your deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                                elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                    your_move_notifier = "Opponent's deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                                    opponent_move_notifier = "Your deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                                else:  # player 2 more than player 1
-                                    your_move_notifier = "YOpponent's deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                                    opponent_move_notifier = "Your deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                                response["winThisTurn"] = True
-                            elif len(game.player1.deck) == 0:
-                                if game.gofirst == game.player1_username:
-                                    # award extra turn
-                                    game.nextturnislast = True
-                                    your_move_notifier = "Warning: Your opponent's next turn is the last turn, due to your deck running out of cards."
-                                    opponent_move_notifier = "Warning: Your opponent's next turn is the last, due to their deck running out of cards."
-                                else:
-                                    # end game and compute winner
-                                    if game.player1.gameDefenceFulfilled() > game.player2.gameDefenceFulfilled():
-                                        your_move_notifier = "Your deck ran out of cards. You win due to having more defence fulfilled!"
-                                        opponent_move_notifier = "Opponent's deck ran out of cards. You lose due to having less defence fulfilled!"
-                                    elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                        your_move_notifier = "Your deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                        opponent_move_notifier = "Opponent's deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                    else:  # player 2 more than player 1
-                                        your_move_notifier = "Your deck ran out of cards. You lose due to having less defence fulfilled!"
-                                        opponent_move_notifier = "Opponent's deck ran out of cards. You win due to having more defence fulfilled!"
-                                    response["winThisTurn"] = True
-                            else:
-                                poppedCard = game.player2.popDeck()
-                                game.player2.addHandCard(poppedCard)  # add the popped card to the hand
-
-                                game.player2.setHandEnablePlayStatus(True)
-                                game.player1.setHandEnablePlayStatus(False)
-                                game.recomputeBlockAndDialogStatus()
-                                game.turn = game.player2_username
-
-                                your_move_notifier = "Opponent's turn."
-                                opponent_move_notifier = f"Your turn. You drew {lookup[poppedCard]}."
-
-                                response["nextTurn"] = True
-
-                        response["hand"] = game.player1.hand
-                        response["discard"] = game.player1.discard
-                        response["cardsLeft"] = len(game.player1.deck)
-                        response["cardPlayed"] = cardPlayed
-                        response["field"] = game.player1.field
-                        response["moveNotifier"] = your_move_notifier
-                        response["canClickEndTurn"] = game.player1_username == game.turn
-
-                        updater = {"uuid": game_id, "username": game.player2_username,
-                                   "hand": game.player1.hand, "discard": game.player1.discard,
-                                   "cardsLeft": len(game.player1.deck), "field": game.player1.field,
-                                   "moveNotifier": opponent_move_notifier}
-                        socketio.emit("update opponent state", updater)
-
-                        updater = {"uuid": game_id, "username": game.player2_username,
-                                   "hand": game.player2.hand, "discard": game.player2.discard,
-                                   "cardsLeft": len(game.player2.deck), "field": game.player2.field,
-                                   "canClickEndTurn": game.turn == game.player2_username}
-                        socketio.emit("update your state", updater)
-
-                        updater = {"uuid": game_id, "username": game.player1_username,
-                                   "hand": game.player2.hand, "discard": game.player2.discard,
-                                   "cardsLeft": len(game.player2.deck), "field": game.player2.field}
-                        socketio.emit("update opponent state", updater)
-
-                        return response
-                elif game.player2_username == request_username:
-                    if (hand_index < 0) or (hand_index >= len(game.player2.hand)):
-                        abort(Response(json.dumps({"Message": "Card Index Out Of Range"}), 422))
-                    else:
-                        cardPlayed = game.player2.hand.pop(hand_index)["name"]
-                        game.player2.discard = [cardPlayed] + game.player2.discard
-
-                        next_turn = len(game.player2.hand) > 7
-
-                        if next_turn:
-                            your_move_notifier = "Please discard cards until you have 7 cards."
-                            opponent_move_notifier = "Waiting for opponent to discard hand cards."
-                        else:
-                            if game.nextturnislast:
-                                if game.player1.gameDefenceFulfilled() > game.player2.gameDefenceFulfilled():
-                                    your_move_notifier = "Opponent's deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                                    opponent_move_notifier = "Your deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                                elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                    your_move_notifier = "Opponent's deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                                    opponent_move_notifier = "Your deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                                else:  # player 2 more than player 1
-                                    your_move_notifier = "YOpponent's deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                                    opponent_move_notifier = "Your deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                                response["winThisTurn"] = True
-                            elif len(game.player1.deck) == 0:
-                                if game.gofirst == game.player1_username:
-                                    # award extra turn
-                                    game.nextturnislast = True
-                                    your_move_notifier = "Warning: Your opponent's next turn is the last turn, due to your deck running out of cards."
-                                    opponent_move_notifier = "Warning: Your opponent's next turn is the last, due to their deck running out of cards."
-                                else:
-                                    # end game and compute winner
-                                    if game.player1.gameDefenceFulfilled() > game.player2.gameDefenceFulfilled():
-                                        your_move_notifier = "Your deck ran out of cards. You win due to having more defence fulfilled!"
-                                        opponent_move_notifier = "Opponent's deck ran out of cards. You lose due to having less defence fulfilled!"
-                                    elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                        your_move_notifier = "Your deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                        opponent_move_notifier = "Opponent's deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                    else:  # player 2 more than player 1
-                                        your_move_notifier = "Your deck ran out of cards. You lose due to having less defence fulfilled!"
-                                        opponent_move_notifier = "Opponent's deck ran out of cards. You win due to having more defence fulfilled!"
-                                    response["winThisTurn"] = True
-                            else:
-                                poppedCard = game.player1.popDeck()
-                                game.player1.addHandCard(poppedCard)  # add the popped card to the hand
-
-                                game.player1.setHandEnablePlayStatus(True)
-                                game.player2.setHandEnablePlayStatus(False)
-                                game.recomputeBlockAndDialogStatus()
-                                game.turn = game.player1_username
-
-                                your_move_notifier = "Opponent's turn."
-                                opponent_move_notifier = f"Your turn. You drew {lookup[poppedCard]}."
-
-                            response["nextTurn"] = True
-
-                        response["hand"] = game.player2.hand
-                        response["discard"] = game.player2.discard
-                        response["cardsLeft"] = len(game.player2.deck)
-                        response["cardPlayed"] = cardPlayed
-                        response["field"] = game.player2.field
-                        response["moveNotifier"] = your_move_notifier
-                        response["canClickEndTurn"] = game.player2_username == game.turn
-
-                        updater = {"uuid": game_id, "username": game.player1_username,
-                                   "hand": game.player2.hand, "discard": game.player2.discard,
-                                   "cardsLeft": len(game.player2.deck), "field": game.player2.field,
-                                   "moveNotifier": opponent_move_notifier}
-                        socketio.emit("update opponent state", updater)
-
-                        updater = {"uuid": game_id, "username": game.player1_username,
-                                   "hand": game.player1.hand, "discard": game.player1.discard,
-                                   "cardsLeft": len(game.player1.deck), "field": game.player1.field,
-                                   "canClickEndTurn": game.player1_username == game.turn}
-                        socketio.emit("update your state", updater)
-
-                        updater = {"uuid": game_id, "username": game.player2_username,
-                                   "hand": game.player1.hand, "discard": game.player1.discard,
-                                   "cardsLeft": len(game.player1.deck), "field": game.player1.field}
-                        socketio.emit("update opponent state", updater)
-
-                        return response
         # else
         abort(Response(json.dumps({"Message": "Cannot Play Hand"}), 404))
 
@@ -1694,169 +738,14 @@ def pass_turn():
     your_username = request.json['username']
     request_username = request.json["request_username"]
     game_id = request.json['game_id']
-    response = {
-        "hand": [],
-        "discard": [],
-        "cardsLeft": 0,
-        "nextTurn": False,
-        "winThisTurn": False
-    }
+    # response = {"hand": [], "discard": [], "cardsLeft": 0, "nextTurn": False, "winThisTurn": False}
+    # response is constructed in the game pass turn function
     if request.method == "POST":
         for i in logged_in:
             keys_equal = secrets.compare_digest(i.login_session_key, sent_login_sesh_key)
             if (i.name == your_username) and keys_equal:
                 game: Game = games[game_id]
-                if game.player1_username == request_username:
-                    next_turn = len(game.player1.hand) > 7
-
-                    if next_turn:
-                        your_move_notifier = "Please discard cards until you have 7 cards."
-                        opponent_move_notifier = "Waiting for opponent to discard hand cards."
-                    else:
-                        if game.nextturnislast:
-                            if game.player1.gameDefenceFulfilled() > game.player2.gameDefenceFulfilled():
-                                your_move_notifier = "Opponent's deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                                opponent_move_notifier = "Your deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                            elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                your_move_notifier = "Opponent's deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                                opponent_move_notifier = "Your deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                            else:  # player 2 more than player 1
-                                your_move_notifier = "YOpponent's deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                                opponent_move_notifier = "Your deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                            response["winThisTurn"] = True
-                        elif len(game.player1.deck) == 0:
-                            if game.gofirst == game.player1_username:
-                                # award extra turn
-                                game.nextturnislast = True
-                                your_move_notifier = "Warning: Your opponent's next turn is the last turn, due to your deck running out of cards."
-                                opponent_move_notifier = "Warning: Your opponent's next turn is the last, due to their deck running out of cards."
-                            else:
-                                # end game and compute winner
-                                if game.player1.gameDefenceFulfilled() > game.player2.gameDefenceFulfilled():
-                                    your_move_notifier = "Your deck ran out of cards. You win due to having more defence fulfilled!"
-                                    opponent_move_notifier = "Opponent's deck ran out of cards. You lose due to having less defence fulfilled!"
-                                elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                    your_move_notifier = "Your deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                    opponent_move_notifier = "Opponent's deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                else:  # player 2 more than player 1
-                                    your_move_notifier = "Your deck ran out of cards. You lose due to having less defence fulfilled!"
-                                    opponent_move_notifier = "Opponent's deck ran out of cards. You win due to having more defence fulfilled!"
-                                response["winThisTurn"] = True
-                        else:
-                            poppedCard = game.player2.popDeck()
-                            game.player2.addHandCard(poppedCard)  # add the popped card to the hand
-
-                            game.player2.setHandEnablePlayStatus(True)
-                            game.player1.setHandEnablePlayStatus(False)
-                            game.recomputeBlockAndDialogStatus()
-                            game.turn = game.player2_username
-
-                            your_move_notifier = "Opponent's turn."
-                            opponent_move_notifier = f"Your turn. You drew {lookup[poppedCard]}."
-
-                        response["nextTurn"] = True
-
-                    response["hand"] = game.player1.hand
-                    response["discard"] = game.player1.discard
-                    response["cardsLeft"] = len(game.player1.deck)
-                    response["field"] = game.player1.field
-                    response["moveNotifier"] = your_move_notifier
-                    response["canClickEndTurn"] = game.player1_username == game.turn
-
-                    updater = {"uuid": game_id, "username": game.player2_username,
-                               "hand": game.player1.hand, "discard": game.player1.discard,
-                               "cardsLeft": len(game.player1.deck), "field": game.player1.field,
-                               "moveNotifier": opponent_move_notifier}
-                    socketio.emit("update opponent state", updater)
-
-                    updater = {"uuid": game_id, "username": game.player2_username,
-                               "hand": game.player2.hand, "discard": game.player2.discard,
-                               "cardsLeft": len(game.player2.deck), "field": game.player2.field,
-                               "canClickEndTurn": game.player2_username == game.turn}
-                    socketio.emit("update your state", updater)
-
-                    updater = {"uuid": game_id, "username": game.player1_username,
-                               "hand": game.player2.hand, "discard": game.player2.discard,
-                               "cardsLeft": len(game.player2.deck), "field": game.player2.field}
-                    socketio.emit("update opponent state", updater)
-
-                    return response
-
-                elif game.player2_username == request_username:
-                    next_turn = len(game.player2.hand) > 7
-
-                    if next_turn:
-                        your_move_notifier = "Please discard cards until you have 7 cards."
-                        opponent_move_notifier = "Waiting for opponent to discard hand cards."
-                    else:
-                        if game.nextturnislast:
-                            if game.player1.gameDefenceFulfilled() > game.player2.gameDefenceFulfilled():
-                                your_move_notifier = "Opponent's deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                                opponent_move_notifier = "Your deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                            elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                your_move_notifier = "Opponent's deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                                opponent_move_notifier = "Your deck ran out of cards 1 turn ago. Tie due to having equal defence fulfilled!"
-                            else:  # player 2 more than player 1
-                                your_move_notifier = "YOpponent's deck ran out of cards 1 turn ago. You lose due to having less defence fulfilled!"
-                                opponent_move_notifier = "Your deck ran out of cards 1 turn ago. You win due to having more defence fulfilled!"
-                            response["winThisTurn"] = True
-                        elif len(game.player1.deck) == 0:
-                            if game.gofirst == game.player1_username:
-                                # award extra turn
-                                game.nextturnislast = True
-                                your_move_notifier = "Warning: Your opponent's next turn is the last turn, due to your deck running out of cards."
-                                opponent_move_notifier = "Warning: Your opponent's next turn is the last, due to their deck running out of cards."
-                            else:
-                                # end game and compute winner
-                                if game.player1.gameDefenceFulfilled() > game.player2.gameDefenceFulfilled():
-                                    your_move_notifier = "Your deck ran out of cards. You win due to having more defence fulfilled!"
-                                    opponent_move_notifier = "Opponent's deck ran out of cards. You lose due to having less defence fulfilled!"
-                                elif game.player1.gameDefenceFulfilled() == game.player2.gameDefenceFulfilled():
-                                    your_move_notifier = "Your deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                    opponent_move_notifier = "Opponent's deck ran out of cards. Tie due to having equal defence fulfilled!"
-                                else:  # player 2 more than player 1
-                                    your_move_notifier = "Your deck ran out of cards. You lose due to having less defence fulfilled!"
-                                    opponent_move_notifier = "Opponent's deck ran out of cards. You win due to having more defence fulfilled!"
-                                response["winThisTurn"] = True
-                        else:
-                            poppedCard = game.player1.popDeck()
-                            game.player1.addHandCard(poppedCard)  # add the popped card to the hand
-
-                            game.player1.setHandEnablePlayStatus(True)
-                            game.player2.setHandEnablePlayStatus(False)
-                            game.recomputeBlockAndDialogStatus()
-                            game.turn = game.player1_username
-
-                            your_move_notifier = "Opponent's turn."
-                            opponent_move_notifier = f"Your turn. You drew {lookup[poppedCard]}."
-
-                        response["nextTurn"] = True
-
-                    response["hand"] = game.player2.hand
-                    response["discard"] = game.player2.discard
-                    response["cardsLeft"] = len(game.player2.deck)
-                    response["field"] = game.player2.field
-                    response["moveNotifier"] = your_move_notifier
-                    response["canClickEndTurn"] = game.player2_username == game.turn
-
-                    updater = {"uuid": game_id, "username": game.player1_username,
-                               "hand": game.player2.hand, "discard": game.player2.discard,
-                               "cardsLeft": len(game.player2.deck), "field": game.player2.field,
-                               "moveNotifier": opponent_move_notifier}
-                    socketio.emit("update opponent state", updater)
-
-                    updater = {"uuid": game_id, "username": game.player1_username,
-                               "hand": game.player1.hand, "discard": game.player1.discard,
-                               "cardsLeft": len(game.player1.deck), "field": game.player1.field,
-                               "canClickEndTurn": game.player1_username == game.turn}
-                    socketio.emit("update your state", updater)
-
-                    updater = {"uuid": game_id, "username": game.player2_username,
-                               "hand": game.player1.hand, "discard": game.player1.discard,
-                               "cardsLeft": len(game.player1.deck), "field": game.player1.field}
-                    socketio.emit("update opponent state", updater)
-
-                    return response
+                return game.pass_turn(socketio, 1 if game.player1_username == request_username else 2)
         # else
         abort(Response(json.dumps({"Message": "Cannot Play Hand"}), 404))
 
@@ -1923,109 +812,7 @@ def game_init():
                     game: Game = games[game_id]
                 except KeyError:
                     return "Not Found"
-                if game.player1_username == your_username or game.player2_username == your_username:
-                    notifier = ""
-                    if game.player1.crisis > game.player2.crisis and game.player1_username == your_username:
-                        if len(game.player1.deck) == 46:  # fresh game, should initialise
-                            for q in range(5):
-                                game.player1.addHandCard(game.player1.popDeck())
-                            game.player1.setHandEnablePlayStatus(True)
-                            game.recomputeBlockAndDialogStatus()
-                            notifier = f"Your crisis has the higher number. You are going first. Drew {lookup[game.player1.hand[-1]['name']]}."
-                        returned = {"username": your_username,
-                                    "hand": game.player1.hand,
-                                    "cardsLeft": len(game.player1.deck),
-                                    "field": game.player1.field,
-                                    "discard": game.player1.discard,
-                                    "crisis": game.player1.crisis,
-                                    "moveNotifier": notifier,
-                                    "uuid": game.internal_id,
-                                    "canClickEndTurn": game.turn == game.player1_username,
-                                    "storage": game.player1.storage}
-                        socketio.emit("update your state", returned)
-                        socketio.emit("update opponent state", {"username": game.player2_username,
-                                                                "cardsLeft": len(game.player1.deck),
-                                                                "field": game.player1.field,
-                                                                "discard": game.player1.discard,
-                                                                "crisis": game.player1.crisis,
-                                                                "uuid": game.internal_id})
-                        return "First"
-                    elif game.player1.crisis < game.player2.crisis and game.player1_username == your_username:
-                        if len(game.player1.deck) == 46:  # fresh game, should initialise
-                            for q in range(5):
-                                game.player1.addHandCard(game.player1.popDeck())
-                            game.player1.setHandEnablePlayStatus(False)
-                            game.recomputeBlockAndDialogStatus()
-                            notifier = f"Your crisis has the lower number.  You are going second."
-                        returned = {"username": your_username,
-                                    "hand": game.player1.hand,
-                                    "cardsLeft": len(game.player1.deck),
-                                    "field": game.player1.field,
-                                    "discard": game.player1.discard,
-                                    "crisis": game.player1.crisis,
-                                    "moveNotifier": notifier,
-                                    "uuid": game.internal_id,
-                                    "canClickEndTurn": game.turn == game.player1_username,
-                                    "storage": game.player1.storage}
-                        socketio.emit("update your state", returned)
-                        socketio.emit("update opponent state", {"username": game.player2_username,
-                                                                "cardsLeft": len(game.player1.deck),
-                                                                "field": game.player1.field,
-                                                                "discard": game.player1.discard,
-                                                                "crisis": game.player1.crisis,
-                                                                "uuid": game.internal_id})
-                        return "Second"
-                    elif game.player1.crisis > game.player2.crisis and game.player2_username == your_username:
-                        if len(game.player2.deck) == 46:  # fresh game, should initialise
-                            for q in range(5):
-                                game.player2.addHandCard(game.player2.popDeck())
-                            game.player2.setHandEnablePlayStatus(False)
-                            game.recomputeBlockAndDialogStatus()
-                            notifier = f"Your crisis has the lower number.  You are going second."
-                        socketio.emit("update your state", {"username": your_username,
-                                                            "hand": game.player2.hand,
-                                                            "cardsLeft": len(game.player2.deck),
-                                                            "field": game.player2.field,
-                                                            "discard": game.player2.discard,
-                                                            "crisis": game.player2.crisis,
-                                                            "moveNotifier": notifier,
-                                                            "uuid": game.internal_id,
-                                                            "canClickEndTurn": game.turn == game.player2_username,
-                                                            "storage": game.player2.storage})
-                        socketio.emit("update opponent state", {"username": game.player1_username,
-                                                                "cardsLeft": len(game.player2.deck),
-                                                                "field": game.player2.field,
-                                                                "discard": game.player2.discard,
-                                                                "crisis": game.player2.crisis,
-                                                                "uuid": game.internal_id})
-                        return "Second"
-                    else:  # player 2 crisis bigger and you are player 2
-                        if len(game.player2.deck) == 46:  # fresh game, should initialise
-                            for q in range(5):
-                                game.player2.addHandCard(game.player2.popDeck())
-                            game.player2.setHandEnablePlayStatus(True)
-                            game.recomputeBlockAndDialogStatus()
-                            notifier = f"Your crisis has the higher number. You are going first. Drew {lookup[game.player2.hand[-1]['name']]}."
-                        socketio.emit("update your state", {"username": your_username,
-                                                            "hand": game.player2.hand,
-                                                            "cardsLeft": len(game.player2.deck),
-                                                            "field": game.player2.field,
-                                                            "discard": game.player2.discard,
-                                                            "crisis": game.player2.crisis,
-                                                            "moveNotifier": notifier,
-                                                            "uuid": game.internal_id,
-                                                            "canClickEndTurn": game.turn == game.player2_username,
-                                                            "storage": game.player2.storage})
-                        socketio.emit("update opponent state", {"username": game.player1_username,
-                                                                "cardsLeft": len(game.player2.deck),
-                                                                "field": game.player2.field,
-                                                                "discard": game.player2.discard,
-                                                                "crisis": game.player2.crisis,
-                                                                "uuid": game.internal_id})
-                        return "First"
-                else:
-                    # todo: spectator mode
-                    return "Not Playing"
+                return game.game_init(socketio, your_username)
         # else
         abort(Response(json.dumps({"Message": "Checking Unavailable"}), 404))
