+++ backend/lists.py
@@ -1,0 +1,99 @@
+# standard deck is a list of the name strings of cards
+STANDARD_DECK: list[str] = [
+    "communitysupport",
+    "communitysupport",
+    "communitysupport",
+    "communitysupport",
+
+    "event-1",
+    "event-2",
+    "event-3",
+    "event-4",
+    "event-5",
+    "event-6",
+    "event-7",
+    "event-8",
+    "event-9",
+    "event-10",
+    "event-11",
+    "event-12",
+
+    "civil-1",
+    "civil-1",
+    "civil-2",
+    "civil-2",
+    "civil-3",
+    "digital-1",
+    "digital-1",
+    "digital-2",
+    "digital-2",
+    "digital-3",
+    "economic-1",
+    "economic-2",
+    "economic-3",
+    "economic-4",
+    "economic-5",
+    "military-1",
+    "military-1",
+    "military-2",
+    "military-3",
+    "military-4",
+    "psychological-1",
+    "psychological-1",
+    "psychological-2",
+    "psychological-2",
+    "psychological-3",
+    "social-1",
+    "social-1",
+    "social-2",
+    "social-2",
+    "social-3",
+]
+
+# select crisis from here
+CRISIS_DECK: list[str] = [
+    "crisis-1",
+    "crisis-2",
+    "crisis-3",
+    "crisis-4",
+    "crisis-5",
+    "crisis-6",
+]
+
+# get name from id string
+LOOKUP: dict[str, str] = {
+    "communitysupport": "Community Support",
+    "event-1": "Racial and Religious Tension",
+    "event-2": "Complaint Culture",
+    "event-3": "Stolen Personal Data",
+    "event-4": "Desensitised Citizens",
+    "event-5": "Disgruntled NSmen",
+    "event-6": "Criminal Collaboration",
+    "event-7": "Blood Bank Runs Dry",
+    "event-8": "Ransomware Attack",
+    "event-9": "Heightened Terrorist Alert",
+    "event-10": "Self Radicalisation",
+    "event-11": "Lone Wolf Attack",
+    "event-12": "Recruiting Desperate People",
+    "civil-1": "Alert Community",
+    "civil-2": "Prepared for Crisis",
+    "civil-3": "Civil Emergency",
+    "digital-1": "Use of Strong Passwords",
+    "digital-2": "Robust Emergency Protocols",
+    "digital-3": "Responsible Social Media Use",
+    "economic-1": "Career Mobility",
+    "economic-2": "Business Resilience",
+    "economic-3": "Financial Aid Schemes",
+    "economic-4": "Increase Surveillance at Maritime Ports",
+    "economic-5": "Skills Upgrading",
+    "military-1": "NSmen on Guard",
+    "military-2": "Increased Security Checks",
+    "military-3": "Naval Convoys",
+    "military-4": "Conduct of Raids",
+    "psychological-1": "Education",
+    "psychological-2": "Attend National Day Celebrations",
+    "psychological-3": "Strong Resolve",
+    "social-1": "Religious Counselling",
+    "social-2": "Neighbourliness",
+    "social-3": "Show of Solidarity"
+}
